third_party/WebKit/LayoutTests/inspector/elements/styles/styles-update-links-expected.txt

Tests that removal of property following its disabling works.

Red text here.
[expanded] 
element.style  { ()

[expanded] 
@media (min-device-width: 1px) (styles-update-links.css:10 -> styles-update-links.css:10:1)
.should-change  { (styles-update-links.css:11 -> styles-update-links.css:11:5)
font-family: monospace;

[expanded] 
.should-change  { (styles-update-links.css:7 -> styles-update-links.css:7:63)
position: relative;
margin: 1em;

[expanded] 
.should-change  { (styles-update-links.css:7 -> styles-update-links.css:7:19)
border: 1px solid black;

[expanded] 
.should-change  { (styles-update-links.css:6 -> styles-update-links.css:6:3)
color: red;

[expanded] 
.left-intact  { (styles-update-links.css:4 -> styles-update-links.css:4:1)
padding: 1em;

[expanded] 
.left-intact  { (styles-update-links.css:1 -> styles-update-links.css:1:1)
box-sizing: border-box;

[expanded] 
div  { (user agent stylesheet)
display: block;




#### AFTER PROPERTY INSERTION ####


[expanded] 
element.style  { ()

[expanded] 
@media (min-device-width: 1px) (styles-update-links.css:12 -> styles-update-links.css:12:1)
.should-change  { (styles-update-links.css:13 -> styles-update-links.css:13:5)
font-family: monospace;

[expanded] 
.should-change  { (styles-update-links.css:9 -> styles-update-links.css:9:47)
position: relative;
margin: 1em;

[expanded] 
.should-change  { (styles-update-links.css:9 -> styles-update-links.css:9:3)
border: 1px solid black;

[expanded] 
.should-change  { (styles-update-links.css:6 -> styles-update-links.css:6:3)
color: red;
/-- overloaded --/ PROPERTY: INSERTED;

[expanded] 
.left-intact  { (styles-update-links.css:4 -> styles-update-links.css:4:1)
padding: 1em;

[expanded] 
.left-intact  { (styles-update-links.css:1 -> styles-update-links.css:1:1)
box-sizing: border-box;

[expanded] 
div  { (user agent stylesheet)
display: block;


4 rule ranges are equal.
4 rule ranges are equal.
3 rule ranges are equal.
4 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
0 rule ranges are equal.


#### AFTER SELECTOR EDIT ####


[expanded] 
element.style  { ()

[expanded] 
@media (min-device-width: 1px) (styles-update-links.css:12 -> styles-update-links.css:12:1)
.should-change  { (styles-update-links.css:13 -> styles-update-links.css:13:5)
font-family: monospace;

[expanded] 
.should-change  { (styles-update-links.css:9 -> styles-update-links.css:9:73)
position: relative;
margin: 1em;

[expanded] 
.should-change, .INSERTED-OTHER-SELECTOR  { (styles-update-links.css:9 -> styles-update-links.css:9:3)
border: 1px solid black;

[expanded] 
.should-change  { (styles-update-links.css:6 -> styles-update-links.css:6:3)
color: red;
/-- overloaded --/ PROPERTY: INSERTED;

[expanded] 
.left-intact  { (styles-update-links.css:4 -> styles-update-links.css:4:1)
padding: 1em;

[expanded] 
.left-intact  { (styles-update-links.css:1 -> styles-update-links.css:1:1)
box-sizing: border-box;

[expanded] 
div  { (user agent stylesheet)
display: block;


4 rule ranges are equal.
4 rule ranges are equal.
4 rule ranges are equal.
4 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
0 rule ranges are equal.


#### AFTER PROPERTY DISABLED ####


[expanded] 
element.style  { ()

[expanded] 
@media (min-device-width: 1px) (styles-update-links.css:12 -> styles-update-links.css:12:1)
.should-change  { (styles-update-links.css:13 -> styles-update-links.css:13:5)
font-family: monospace;

[expanded] 
.should-change  { (styles-update-links.css:9 -> styles-update-links.css:9:79)
position: relative;
margin: 1em;

[expanded] 
.should-change, .INSERTED-OTHER-SELECTOR  { (styles-update-links.css:9 -> styles-update-links.css:9:3)
/-- overloaded --/ /-- disabled --/ /* border: 1px solid black; */

[expanded] 
.should-change  { (styles-update-links.css:6 -> styles-update-links.css:6:3)
color: red;
/-- overloaded --/ PROPERTY: INSERTED;

[expanded] 
.left-intact  { (styles-update-links.css:4 -> styles-update-links.css:4:1)
padding: 1em;

[expanded] 
.left-intact  { (styles-update-links.css:1 -> styles-update-links.css:1:1)
box-sizing: border-box;

[expanded] 
div  { (user agent stylesheet)
display: block;


4 rule ranges are equal.
4 rule ranges are equal.
4 rule ranges are equal.
4 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
0 rule ranges are equal.
